+++ ui/cypress/tests/adapter/machineDataSimulator.smoke.spec.ts
@@ -18,72 +18,19 @@
 
 import { ConnectUtils } from '../../support/utils/connect/ConnectUtils';
 import { AdapterBuilder } from '../../support/builder/AdapterBuilder';
-import { PipelineBuilder } from '../../support/builder/PipelineBuilder';
-import { PipelineElementBuilder } from '../../support/builder/PipelineElementBuilder';
-import { PipelineUtils } from '../../support/utils/PipelineUtils';
-import { UserUtils } from '../../support/utils/UserUtils';
-
-const adapterName = 'simulator';
 
 describe('Test Machine Data Simulator Adapter', () => {
     beforeEach('Setup Test', () => {
         cy.initStreamPipesTest();
     });
 
-    const adapterInput = AdapterBuilder.create('Machine_Data_Simulator')
-        .setName(adapterName)
-        .addInput('input', 'wait-time-ms', '1000')
-        .build();
+    it('Perform Test', () => {
+        const adapterInput = AdapterBuilder.create('Machine_Data_Simulator')
+            .setName('Machine Data Simulator Test')
+            .addInput('input', 'wait-time-ms', '1000')
+            .build();
 
-    it('Test Basic Delete Adapter', () => {
         ConnectUtils.testAdapter(adapterInput);
         ConnectUtils.deleteAdapter();
-    });
-
-    const pipelineInput = PipelineBuilder.create('Pipeline Test')
-        .addSource(adapterName)
-        .addProcessingElement(
-            PipelineElementBuilder.create('field_renamer')
-                .addInput('drop-down', 'convert-property', 'timestamp')
-                .addInput('input', 'field-name', 't')
-                .build(),
-        )
-        .addSink(
-            PipelineElementBuilder.create('data_lake')
-                .addInput('input', 'db_measurement', 'demo')
-                .build(),
-        )
-        .build();
-
-    it('Test Delete Adapter and Associated Pipelines', () => {
-        UserUtils.addUser(UserUtils.adapterAndPipelineAdminUser);
-        cy.switchUser(UserUtils.adapterAndPipelineAdminUser);
-        ConnectUtils.testAdapter(adapterInput);
-        PipelineUtils.addPipeline(pipelineInput);
-        PipelineUtils.addPipeline(pipelineInput);
-        ConnectUtils.deleteAdapterAndAssociatedPipelines();
-    });
-
-    it('Test Admin Should Be Able to Delete Adapter and Not Owned Associated Pipelines', () => {
-        // Let the user create the adapter and the pipeline
-        // For some reason, I have to first go to a page before switching user
-        UserUtils.goToUserConfiguration();
-        cy.switchUser(UserUtils.adapterAndPipelineAdminUser);
-        ConnectUtils.testAdapter(adapterInput);
-        PipelineUtils.addPipeline(pipelineInput);
-        PipelineUtils.addPipeline(pipelineInput);
-        // Then let the admin delete them
-        cy.switchUser(UserUtils.adminUser);
-        ConnectUtils.deleteAdapterAndAssociatedPipelines(true);
-    });
-
-    it('Test Delete Adapter and Associated Pipelines Permission Denied', () => {
-        // Let the admin create the adapter and the pipeline
-        ConnectUtils.testAdapter(adapterInput);
-        PipelineUtils.addPipeline(pipelineInput);
-        PipelineUtils.addPipeline(pipelineInput);
-        // Then the user shouldn't be able to delete them
-        cy.switchUser(UserUtils.adapterAndPipelineAdminUser);
-        ConnectUtils.deleteAdapterAndAssociatedPipelinesPermissionDenied();
     });
 });
